Progress keeping - doesn't build
Building better genesis and exchange protocol.

+++ sql-interface/sqliteodbc-0.9996-hschain/sqlite3odbc.h
@@ -157,7 +157,7 @@ typedef struct dbc {
 
     // hschain additions.
     int hschain_node_sockfd; /**< non-negative value means we have a connection */
-    int current_height;
+    int64_t current_height;
     int height_increment;
     uint8_t public_key[256]; /**< just a string, BTW. May switch to binary representation someday. */
     uint8_t private_key[256];
